refactor: slim ui to a pure design system (stage 5/N)
`packages/ui` was mixing shared primitives (used by web + desktop) with
app-only AI-chat components that only the desktop renderer uses and that
dragged heavy deps into the shared package. Move those into the renderer:

- ui/components/ai-elements/* + chat-message + file-thumbnail + input-group
  → apps/desktop/src/renderer/ai-elements/
- migrate their deps ui → desktop (streamdown + @streamdown/*, pdfjs-dist,
  use-stick-to-bottom, nanoid; framer-motion/motion/cva shared → also on
  desktop). ui drops 8 deps (24 → 16).
- move the streamdown markdown-render CSS out of ui's shared globals.css into
  the renderer's styles.css (web doesn't render chat markdown), and drop the
  dead `packages/app` @source glob
- trim the moved components' now-dead library-style exports

`ui` is now purely the shared design system (primitives + the Three.js orb +
motion helpers) that both apps/web and the desktop renderer consume. All gates
green.
