+++ apps/desktop/package.json
@@ -48,9 +48,15 @@
     "@repo/features": "workspace:^",
     "@repo/ui": "workspace:^",
     "@sinclair/typebox": "catalog:",
+    "@streamdown/cjk": "^1.0.3",
+    "@streamdown/code": "^1.1.1",
+    "@streamdown/math": "^1.0.2",
+    "@streamdown/mermaid": "^1.0.2",
     "ai": "^6.0.191",
+    "class-variance-authority": "^0.7.1",
     "d3-force": "^3.0.0",
     "electron-updater": "^6.8.3",
+    "framer-motion": "^12.42.1",
     "katex": "^0.16.22",
     "lowlight": "^3.3.0",
     "lucide-react": "catalog:",
@@ -59,6 +65,9 @@
     "mdast-util-to-markdown": "^2.1.2",
     "mermaid": "^11.16.0",
     "micromark-extension-mdx": "^2.1.0",
+    "motion": "^12.40.0",
+    "nanoid": "^5.1.11",
+    "pdfjs-dist": "^6.1.200",
     "platejs": "^53.2.1",
     "react": "catalog:",
     "react-day-picker": "^9.14.0",
@@ -71,7 +80,9 @@
     "remark-math": "^6.0.0",
     "remark-parse": "^11.0.0",
     "sherpa-onnx-node": "^1.10.41",
+    "streamdown": "^2.5.0",
     "unified": "^11.0.5",
+    "use-stick-to-bottom": "^1.1.4",
     "zustand": "^5.0.13"
   },
   "devDependencies": {
