+++ apps/desktop/src/renderer/ai-elements/attachments.tsx
