+++ apps/desktop/src/renderer/ai-elements/conversation.tsx
