+++ apps/desktop/src/renderer/ai-elements/input-group.tsx
