+++ apps/desktop/src/renderer/ai-elements/shimmer.tsx
