+++ apps/desktop/src/renderer/delegation/delegation-dock.tsx
@@ -17,7 +17,7 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
 } from "@repo/ui/components/alert-dialog";
-import { Response } from "@repo/ui/components/ai-elements/response";
+import { Response } from "@renderer/ai-elements/response";
 import { Button } from "@repo/ui/components/button";
 import { toast } from "@repo/ui/components/sonner";
 import { cn } from "@repo/ui/lib/utils";
