+++ packages/ui/package.json
@@ -24,26 +24,17 @@
     "@radix-ui/react-slot": "^1.3.0",
     "@radix-ui/react-tooltip": "^1.2.10",
     "@react-three/fiber": "^9.6.1",
-    "@streamdown/cjk": "^1.0.3",
-    "@streamdown/code": "^1.1.1",
-    "@streamdown/math": "^1.0.2",
-    "@streamdown/mermaid": "^1.0.2",
-    "ai": "^6.0.191",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.1.1",
     "framer-motion": "^12.42.1",
     "lucide-react": "catalog:",
     "motion": "^12.40.0",
-    "nanoid": "^5.1.11",
-    "pdfjs-dist": "^6.1.200",
     "react-dom": "catalog:",
     "sonner": "^2.0.7",
-    "streamdown": "^2.5.0",
     "tailwind-merge": "^3.6.0",
     "three": "^0.184.0",
-    "tw-animate-css": "^1.4.0",
-    "use-stick-to-bottom": "^1.1.4"
+    "tw-animate-css": "^1.4.0"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "catalog:",
