+++ packages/ui/src/styles/globals.css
@@ -1,11 +1,9 @@
 @import "@fontsource-variable/inter";
 @import "tailwindcss";
-@import "streamdown/styles.css";
 @import "tw-animate-css";
 @import "shadcn/tailwind.css";
 
 @source "../../../../apps/**/*.{ts,tsx}";
-@source "../../../app/src/**/*.{ts,tsx}";
 @source "../**/*.{ts,tsx}";
 
 @custom-variant dark (&:is(.dark *));
